LayoutTests/fast/regions/get-regions-by-content.html

<!doctype html>
<html>
    <head>
        <script src="../../fast/js/resources/js-test-pre.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            body { font-family: monospace; }
            #article { -webkit-flow-into: flow; }
            .border { border: 1px solid red; }
            #region { -webkit-flow-from: flow; width: 250px; height: 50px; position: absolute; left: 10px; top: 200px; }

            #article2 { -webkit-flow-into: flow2; }
            #region21 { -webkit-flow-from: flow2; width: 250px; height: 50px; position: absolute; left: 20px; top: 300px;}
            #region22 { -webkit-flow-from: flow2; width: 250px; height: 50px; position: absolute; left: 20px; top: 350px;}
        </style>
    </head>
    <body>
        <div id="article">
            <p id="p1">Text flown into region.</p>
        </div>
        <div id="region" class="border"></div>

        <div id="article2">
            <p id="p2">Text flown in two regions. Text flown in two regions. Text flown in two regions.</p>
        </div>
        <div id="region21" class="border"></div>
        <div id="region22" class="border"></div>
        <script>
            if (window.testRunner)
                testRunner.dumpAsText();

            description("Test for 90759: [CSSRegions] Rename NamedFlow::getRegionsByContentNode to NamedFlow::getRegionsByContent.");

            var namedFlow = getFlowByName("flow");
            var targetNode = document.getElementById("p1");
            var regionList = namedFlow.getRegionsByContent(targetNode);

            shouldEvaluateTo("regionList.length", 1);
            shouldBeEqualToString("regionList.item(0).id", "region");

            namedFlow = getFlowByName("flow2");
            // Test that the p2 element is flown into 2 regions
            targetNode = document.getElementById("p2");
            regionList = namedFlow.getRegionsByContent(targetNode);

            shouldEvaluateTo("regionList.length", 2);
            shouldBeEqualToString("regionList.item(0).id", "region21");
            shouldBeEqualToString("regionList.item(1).id", "region22");

            // Test that p1 element is not in flow2
            targetNode = document.getElementById("p1");
            regionList = namedFlow.getRegionsByContent(targetNode);
            shouldEvaluateTo("regionList.length", 0);

            function hideRegionsAndFlows() {
                document.getElementById("article").style.visibility = "hidden";
                document.getElementById("article2").style.visibility = "hidden";
                document.getElementById("region").style.visibility = "hidden";
                document.getElementById("region21").style.visibility = "hidden";
                document.getElementById("region22").style.visibility = "hidden";
            }

            hideRegionsAndFlows();
        </script>
        <script src="../../fast/js/resources/js-test-post.js"></script>
    </body>
</html>
